refactor(identity): move artifact_identifier to rac.core.identity [roadmap:v0.7.5]
